+++ Raytracer/Geometry/Sphere.cs
@@ -1,15 +1,13 @@
 ﻿using Raytracer.BaseClasses;
 using Raytracer.Materials;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Raytracer.Geometry
 {
     public class Sphere : Hitable
     {
         private Vec3 _center;
-        private float _radius;
+        private double _radius;
         private Material _material;
 
         #region Constructors
@@ -18,7 +16,7 @@ namespace Raytracer.Geometry
 
         }
 
-        public Sphere(Vec3 cen, float r, Material mat)
+        public Sphere(Vec3 cen, double r, Material mat)
         {
             _center = cen;
             _radius = r;
@@ -26,18 +24,18 @@ namespace Raytracer.Geometry
         }
         #endregion
 
-        public override bool Hit(Ray r, float tMin, float tMax, out HitRecord rec)
+        public override bool Hit(Ray r, double tMin, double tMax, out HitRecord rec)
         {
             Vec3 oc = r.Origin() - _center;
-            float a = Vec3.Dot(r.Direction(), r.Direction());
-            float b = Vec3.Dot(oc, r.Direction());
-            float c = Vec3.Dot(oc, oc) - _radius * _radius;
+            double a = Vec3.Dot(r.Direction(), r.Direction());
+            double b = Vec3.Dot(oc, r.Direction());
+            double c = Vec3.Dot(oc, oc) - _radius * _radius;
 
-            float discriminant = b * b - a * c;
+            double discriminant = b * b - a * c;
 
             if (discriminant > 0)
             {
-                float temp = (-b - (float)Math.Sqrt(b * b - a * c)) / a;
+                double temp = (-b - Math.Sqrt(b * b - a * c)) / a;
                 if (temp < tMax && temp > tMin)
                 {
                     rec.T = temp;
@@ -47,7 +45,7 @@ namespace Raytracer.Geometry
                     return true;
                 }
 
-                temp = (-b + (float)Math.Sqrt(b * b - a * c)) / a;
+                temp = (-b + Math.Sqrt(b * b - a * c)) / a;
                 if (temp < tMax && temp > tMin)
                 {
                     rec.T = temp;
@@ -58,7 +56,7 @@ namespace Raytracer.Geometry
                 }
             }
 
-            rec.T = 0.0F;
+            rec.T = 0.0;
             rec.P = null;
             rec.Normal = null;
             rec.Material = null;
